+++ demoproject/demoproject/views.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from django.shortcuts import render_to_response
+from django.shortcuts import render
 #from django.template.context import RequestContext
 
 
@@ -7,4 +7,4 @@ def home(request):
     """
     home page
     """
-    return render_to_response('home.html')
+    return render(request, 'home.html')