+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 * cleanups and refactoring
 * faster queries
+* add Django 3.0 support
 
 0.14.0 (2020-01-28)
 ------------------